renaming admin menu because It was annoying me
