+++ kanidmd_web_ui/src/components/admin_menu.rs
