+++ kanidmd_web_ui/src/components/mod.rs
@@ -4,7 +4,7 @@ use yew::Html;
 pub mod admin_accounts;
 pub mod admin_groups;
 pub mod admin_oauth2;
-pub mod adminmenu;
+pub mod admin_menu;
 pub mod change_unix_password;
 
 /// creates the "Kanidm is alpha" banner
